Implement simple PageData persistance and unpersistence
